# [WIP] Fast path for function tp$call()