This change adds a couple of fast paths to the `tp$call` method of `Sk.builtin.func`. Although the don't break any of the tests, they also make no difference to the amount of time it takes to run the unit tests on my machine.

If it doesn't bring any benefit, we shouldn't carry the extra complexity, but I'm putting this out as a WIP so that @rixner and others with workloads that were slowed down a lot by the 2->3 migration can try it out.

I am inclined to interpret this as evidence that the `tp$call` function was not the bottleneck we were supposing it to be...