@rixner 2019-08-23
@meredydd I have been swamped for the past month or so.  It is definitely on my todo list to come up with a benchmark I can release that demonstrates the slowdowns I am seeing.  I really don't think the tests serve as good performance benchmarks.  I think they are dominated by compile time.

I will run this and report back as soon as I can, but the semester starts for me on Monday and I have countless logistics to deal with at this point in time.

Regardless, I really appreciate the effort to try to get the performance back!  Thanks!