@meredydd 2019-08-23
No problem. I'm happy to leave this PR open until you have time to take a swing at it!