@albertjan 2019-08-25
I've run [`pystone.py`](https://github.com/python/cpython/blob/2.7/Lib/test/pystone.py) through this branch. There's a small performance gain. Results below: 

## Before

```
Pystone(1.1) time for 50000 passes = 5.184000015258789
This machine benchmarks at 9645.061700005408 pystones/second
-----
Run time: 5.611s

Pystone(1.1) time for 50000 passes = 5.021999835968018
This machine benchmarks at 9956.193077087632 pystones/second
-----
Run time: 5.428s

Pystone(1.1) time for 50000 passes = 5.148000001907349
This machine benchmarks at 9712.5097089112 pystones/second
-----
Run time: 5.561s
```

## After

```
Pystone(1.1) time for 50000 passes = 4.926999807357788
This machine benchmarks at 10148.163579249986 pystones/second
-----
Run time: 5.324s

Pystone(1.1) time for 50000 passes = 4.92799973487854
This machine benchmarks at 10146.104441954145 pystones/second
-----
Run time: 5.325s

Pystone(1.1) time for 50000 passes = 5.1600000858306885
This machine benchmarks at 9689.922319439398 pystones/second
-----
Run time: 5.553s

Pystone(1.1) time for 50000 passes = 4.995999813079834
This machine benchmarks at 10008.006779563308 pystones/second
-----
Run time: 5.39s