@rixner 2020-01-21
When I run this in node, I don't see any measurable benefits.  When I run it in Chrome, I see maybe a 5% speedup?  This is on some code that I know exercises the fast path a significant amount.

I can't decide if what I'm seeing in Chrome is statistically significant or just my imagination (as the numbers bounce around a lot as usual).  It's not a large difference, but it does seem like the fast path numbers are consistently better than not having it.

The question is whether the small difference warrants the small increase in complexity.  I think I come down in favor of yes?

Regardless, I don't think we can argue that ```tp$call``` is not the bottleneck.  It still shows up at the very top of the profile output (both with and without this fast path).  It may or may not be the bottleneck, but I don't think this pull request is conclusive evidence that it is not.

Node: 8.10
Chrome: 79.0.3945.130