@meredydd 2020-02-01
This PR is superseded by #1009